chore: move pages to pages folder..
