+++ frontend/components/Pages/MainPage/sections/RewardsSection/StakingRewardsThisEpoch.tsx
