+++ frontend/components/Pages/ManageStakingPage/StakingContractSection/CantMigrateAlert.tsx
