+++ frontend/components/Pages/ManageStakingPage/StakingContractSection/StakingContractDetails.tsx
