+++ frontend/components/Pages/ManageStakingPage/StakingContractSection/index.tsx
