+++ frontend/components/Pages/SettingsPage/AddBackupWalletPage.tsx
@@ -8,8 +8,8 @@ import { SettingsScreen } from '@/enums/SettingsScreen';
 import { useBalance } from '@/hooks/useBalance';
 import { useSettings } from '@/hooks/useSettings';
 
-import { CardTitle } from '../Card/CardTitle';
-import { CardFlex } from '../styled/CardFlex';
+import { CardTitle } from '../../Card/CardTitle';
+import { CardFlex } from '../../styled/CardFlex';
 
 export const AddBackupWalletPage = () => {
   const { eoaBalance } = useBalance();
