+++ frontend/components/Pages/YourWalletPage/styles.ts
